chore: moved landing page, loading page and metrics page to views folder, and renamed pages  accordingly
