+++ commitment/imports/ui/views/MetricsView.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import MetricsTabs from "./components/metrics/MetricsTab";
-import TopBar from "./components/metrics/TopBar";
+import MetricsTabs from "../components/metrics/MetricsTab";
+import TopBar from "../components/metrics/TopBar";
 
 const MetricsPage: React.FC = () => (
     <div className="m-0 scroll-smooth flex flex-col">
